requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
rich>=13.7.0,<14.0.0
pydantic>=2.0.0
reportlab>=4.0.0
streamlit>=1.35.0
plotly>=5.0.0
